+++ openmls/src/group/mls_group/builder.rs
@@ -5,8 +5,8 @@ use crate::{
     extensions::ExternalSendersExtension,
     group::{
         config::CryptoConfig, public_group::errors::PublicGroupBuildError, CoreGroup,
-        CoreGroupBuildError, CoreGroupConfig, GroupId, MlsGroupPattern, MlsGroupPatternBuilder,
-        NewGroupError, ProposalStore, WireFormatPolicy,
+        CoreGroupBuildError, CoreGroupConfig, GroupId, MlsGroupCreateConfig,
+        MlsGroupCreateConfigBuilder, NewGroupError, ProposalStore, WireFormatPolicy,
     },
     prelude::{LibraryError, Lifetime, SenderRatchetConfiguration},
 };
@@ -16,7 +16,7 @@ use super::{InnerState, MlsGroup, MlsGroupState};
 #[derive(Default)]
 pub struct MlsGroupBuilder {
     group_id: Option<GroupId>,
-    mls_group_pattern_builder: MlsGroupPatternBuilder,
+    mls_group_create_config_builder: MlsGroupCreateConfigBuilder,
 }
 
 impl MlsGroupBuilder {
@@ -24,11 +24,13 @@ impl MlsGroupBuilder {
         Self::default()
     }
 
+    /// Sets the group ID of the [`MlsGroup`].
     pub fn with_group_id(mut self, group_id: GroupId) -> Self {
         self.group_id = Some(group_id);
         self
     }
 
+    /// Build a new group as configured by this builder.
     pub fn build<KeyStore: OpenMlsKeyStore>(
         self,
         provider: &impl OpenMlsProvider<KeyStoreProvider = KeyStore>,
@@ -40,7 +42,7 @@ impl MlsGroupBuilder {
 
     /// Build a new group with the given group ID.
     ///
-    /// If an [`MlsGroupPattern`] is provided, it will be used to configure the
+    /// If an [`MlsGroupCreateConfig`] is provided, it will be used to configure the
     /// group. Otherwise, the internal builder is used to build one with the
     /// parameters set on this builder.
     pub(super) fn build_internal<KeyStore: OpenMlsKeyStore>(
@@ -48,30 +50,30 @@ impl MlsGroupBuilder {
         provider: &impl OpenMlsProvider<KeyStoreProvider = KeyStore>,
         signer: &impl Signer,
         credential_with_key: CredentialWithKey,
-        mls_group_pattern_option: Option<MlsGroupPattern>,
+        mls_group_create_config_option: Option<MlsGroupCreateConfig>,
     ) -> Result<MlsGroup, NewGroupError<KeyStore::Error>> {
-        let mls_group_pattern =
-            mls_group_pattern_option.unwrap_or_else(|| self.mls_group_pattern_builder.build());
+        let mls_group_create_config = mls_group_create_config_option
+            .unwrap_or_else(|| self.mls_group_create_config_builder.build());
         let group_id = self
             .group_id
             .unwrap_or_else(|| GroupId::random(provider.rand()));
         // TODO #751
         let group_config = CoreGroupConfig {
-            add_ratchet_tree_extension: mls_group_pattern
+            add_ratchet_tree_extension: mls_group_create_config
                 .mls_group_config
                 .use_ratchet_tree_extension,
         };
 
         let mut group = CoreGroup::builder(
             group_id,
-            mls_group_pattern.crypto_config,
+            mls_group_create_config.crypto_config,
             credential_with_key,
         )
         .with_config(group_config)
-        .with_required_capabilities(mls_group_pattern.required_capabilities.clone())
-        .with_external_senders(mls_group_pattern.external_senders.clone())
-        .with_max_past_epoch_secrets(mls_group_pattern.mls_group_config.max_past_epochs)
-        .with_lifetime(*mls_group_pattern.lifetime())
+        .with_required_capabilities(mls_group_create_config.required_capabilities.clone())
+        .with_external_senders(mls_group_create_config.external_senders.clone())
+        .with_max_past_epoch_secrets(mls_group_create_config.mls_group_config.max_past_epochs)
+        .with_lifetime(*mls_group_create_config.lifetime())
         .build(provider, signer)
         .map_err(|e| match e {
             CoreGroupBuildError::LibraryError(e) => e.into(),
@@ -100,7 +102,7 @@ impl MlsGroupBuilder {
             .add(group.context().epoch(), resumption_psk.clone());
 
         let mls_group = MlsGroup {
-            mls_group_config: mls_group_pattern.mls_group_config.clone(),
+            mls_group_config: mls_group_create_config.mls_group_config.clone(),
             group,
             proposal_store: ProposalStore::new(),
             own_leaf_nodes: vec![],
@@ -112,19 +114,21 @@ impl MlsGroupBuilder {
         Ok(mls_group)
     }
 
-    // MlsGroupPatternBuilder options
+    // MlsGroupCreateConfigBuilder options
 
     /// Sets the `wire_format` property of the MlsGroup.
     pub fn with_wire_format_policy(mut self, wire_format_policy: WireFormatPolicy) -> Self {
-        self.mls_group_pattern_builder = self
-            .mls_group_pattern_builder
+        self.mls_group_create_config_builder = self
+            .mls_group_create_config_builder
             .wire_format_policy(wire_format_policy);
         self
     }
 
     /// Sets the `padding_size` property of the MlsGroup.
     pub fn padding_size(mut self, padding_size: usize) -> Self {
-        self.mls_group_pattern_builder = self.mls_group_pattern_builder.padding_size(padding_size);
+        self.mls_group_create_config_builder = self
+            .mls_group_create_config_builder
+            .padding_size(padding_size);
         self
     }
 
@@ -139,24 +143,24 @@ impl MlsGroupBuilder {
     /// the same epoch in which they were generated. The number for `max_epochs` should be
     /// as low as possible.
     pub fn max_past_epochs(mut self, max_past_epochs: usize) -> Self {
-        self.mls_group_pattern_builder = self
-            .mls_group_pattern_builder
+        self.mls_group_create_config_builder = self
+            .mls_group_create_config_builder
             .max_past_epochs(max_past_epochs);
         self
     }
 
     /// Sets the `number_of_resumption_psks` property of the MlsGroup.
     pub fn number_of_resumption_psks(mut self, number_of_resumption_psks: usize) -> Self {
-        self.mls_group_pattern_builder = self
-            .mls_group_pattern_builder
+        self.mls_group_create_config_builder = self
+            .mls_group_create_config_builder
             .number_of_resumption_psks(number_of_resumption_psks);
         self
     }
 
     /// Sets the `use_ratchet_tree_extension` property of the MlsGroup.
     pub fn use_ratchet_tree_extension(mut self, use_ratchet_tree_extension: bool) -> Self {
-        self.mls_group_pattern_builder = self
-            .mls_group_pattern_builder
+        self.mls_group_create_config_builder = self
+            .mls_group_create_config_builder
             .use_ratchet_tree_extension(use_ratchet_tree_extension);
         self
     }
@@ -167,28 +171,30 @@ impl MlsGroupBuilder {
         mut self,
         sender_ratchet_configuration: SenderRatchetConfiguration,
     ) -> Self {
-        self.mls_group_pattern_builder = self
-            .mls_group_pattern_builder
+        self.mls_group_create_config_builder = self
+            .mls_group_create_config_builder
             .sender_ratchet_configuration(sender_ratchet_configuration);
         self
     }
 
     /// Sets the `lifetime` of the group creator's leaf.
     pub fn lifetime(mut self, lifetime: Lifetime) -> Self {
-        self.mls_group_pattern_builder = self.mls_group_pattern_builder.lifetime(lifetime);
+        self.mls_group_create_config_builder =
+            self.mls_group_create_config_builder.lifetime(lifetime);
         self
     }
 
     /// Sets the `crypto_config` of the MlsGroup.
     pub fn crypto_config(mut self, config: CryptoConfig) -> Self {
-        self.mls_group_pattern_builder = self.mls_group_pattern_builder.crypto_config(config);
+        self.mls_group_create_config_builder =
+            self.mls_group_create_config_builder.crypto_config(config);
         self
     }
 
     /// Sets the `external_senders` property of the MlsGroup.
     pub fn external_senders(mut self, external_senders: ExternalSendersExtension) -> Self {
-        self.mls_group_pattern_builder = self
-            .mls_group_pattern_builder
+        self.mls_group_create_config_builder = self
+            .mls_group_create_config_builder
             .external_senders(external_senders);
         self
     }