+++ openmls/src/group/mls_group/creation.rs
@@ -30,14 +30,14 @@ impl MlsGroup {
     pub fn new<KeyStore: OpenMlsKeyStore>(
         provider: &impl OpenMlsProvider<KeyStoreProvider = KeyStore>,
         signer: &impl Signer,
-        mls_group_pattern: &MlsGroupPattern,
+        mls_group_create_config: &MlsGroupCreateConfig,
         credential_with_key: CredentialWithKey,
     ) -> Result<Self, NewGroupError<KeyStore::Error>> {
         MlsGroupBuilder::new().build_internal(
             provider,
             signer,
             credential_with_key,
-            Some(mls_group_pattern.clone()),
+            Some(mls_group_create_config.clone()),
         )
     }
 
@@ -46,7 +46,7 @@ impl MlsGroup {
     pub fn new_with_group_id<KeyStore: OpenMlsKeyStore>(
         provider: &impl OpenMlsProvider<KeyStoreProvider = KeyStore>,
         signer: &impl Signer,
-        mls_group_pattern: &MlsGroupPattern,
+        mls_group_create_config: &MlsGroupCreateConfig,
         group_id: GroupId,
         credential_with_key: CredentialWithKey,
     ) -> Result<Self, NewGroupError<KeyStore::Error>> {
@@ -56,7 +56,7 @@ impl MlsGroup {
                 provider,
                 signer,
                 credential_with_key,
-                Some(mls_group_pattern.clone()),
+                Some(mls_group_create_config.clone()),
             )
     }
 
@@ -66,7 +66,7 @@ impl MlsGroup {
     // TODO: #1326 This should take an MlsMessage rather than a Welcome message.
     pub fn new_from_welcome<KeyStore: OpenMlsKeyStore>(
         provider: &impl OpenMlsProvider<KeyStoreProvider = KeyStore>,
-        mls_group_config: &MlsGroupConfig,
+        mls_group_config: &MlsGroupJoinConfig,
         welcome: Welcome,
         ratchet_tree: Option<RatchetTreeIn>,
     ) -> Result<Self, WelcomeError<KeyStore::Error>> {
@@ -146,7 +146,7 @@ impl MlsGroup {
         signer: &impl Signer,
         ratchet_tree: Option<RatchetTreeIn>,
         verifiable_group_info: VerifiableGroupInfo,
-        mls_group_config: &MlsGroupConfig,
+        mls_group_config: &MlsGroupJoinConfig,
         aad: &[u8],
         credential_with_key: CredentialWithKey,
     ) -> Result<(Self, MlsMessageOut, Option<GroupInfo>), ExternalCommitError> {