+++ openmls/src/group/tests/external_add_proposal.rs
@@ -36,7 +36,7 @@ fn new_test_group(
         generate_credential_with_key(identity.into(), ciphersuite.signature_algorithm(), provider);
 
     // Define the MlsGroup configuration
-    let mls_group_pattern = MlsGroupPattern::builder()
+    let mls_group_create_config = MlsGroupCreateConfig::builder()
         .wire_format_policy(wire_format_policy)
         .crypto_config(CryptoConfig::with_default_version(ciphersuite))
         .build();
@@ -45,7 +45,7 @@ fn new_test_group(
         MlsGroup::new_with_group_id(
             provider,
             &credential_with_keys.signer,
-            &mls_group_pattern,
+            &mls_group_create_config,
             group_id,
             credential_with_keys.credential_with_key.clone(),
         )
@@ -83,7 +83,7 @@ fn validation_test_setup(
         .expect("error merging pending commit");
 
     // Define the MlsGroup configuration
-    let mls_group_config = MlsGroupConfig::builder()
+    let mls_group_config = MlsGroupJoinConfig::builder()
         .wire_format_policy(wire_format_policy)
         .build();
 
@@ -193,7 +193,9 @@ fn external_add_proposal_should_succeed(ciphersuite: Ciphersuite, provider: &imp
         assert_eq!(bob_group.members().count(), 3);
 
         // Finally, Charlie can join with the Welcome
-        let mls_group_config = MlsGroupConfig::builder().wire_format_policy(policy).build();
+        let mls_group_config = MlsGroupJoinConfig::builder()
+            .wire_format_policy(policy)
+            .build();
         let charlie_group = MlsGroup::new_from_welcome(
             provider,
             &mls_group_config,