+++ openmls/tests/test_external_commit.rs
@@ -9,7 +9,7 @@ fn create_alice_group(
     provider: &impl OpenMlsProvider,
     use_ratchet_tree_extension: bool,
 ) -> (MlsGroup, CredentialWithKey, SignatureKeyPair) {
-    let group_config = MlsGroupPattern::builder()
+    let group_config = MlsGroupCreateConfig::builder()
         .use_ratchet_tree_extension(use_ratchet_tree_extension)
         .crypto_config(CryptoConfig::with_default_version(ciphersuite))
         .build();
@@ -88,7 +88,7 @@ fn test_external_commit(ciphersuite: Ciphersuite, provider: &impl OpenMlsProvide
             &bob_signature_keys,
             None,
             verifiable_group_info,
-            &MlsGroupConfig::default(),
+            &MlsGroupJoinConfig::default(),
             b"",
             bob_credential,
         )
@@ -109,7 +109,7 @@ fn test_external_commit(ciphersuite: Ciphersuite, provider: &impl OpenMlsProvide
             &bob_signature_keys,
             None,
             verifiable_group_info_broken,
-            &MlsGroupConfig::default(),
+            &MlsGroupJoinConfig::default(),
             b"",
             bob_credential,
         )
@@ -151,7 +151,7 @@ fn test_group_info(ciphersuite: Ciphersuite, provider: &impl OpenMlsProvider) {
         &bob_signature_keys,
         None,
         verifiable_group_info,
-        &MlsGroupConfig::default(),
+        &MlsGroupJoinConfig::default(),
         b"",
         bob_credential,
     )
@@ -202,7 +202,7 @@ fn test_group_info(ciphersuite: Ciphersuite, provider: &impl OpenMlsProvider) {
         &bob_signature_keys,
         None,
         verifiable_group_info,
-        &MlsGroupConfig::default(),
+        &MlsGroupJoinConfig::default(),
         b"",
         bob_credential,
     )